requirements.txt
tensorflow-cpu==2.10.0
numpy==1.23.3
